+++ src/loggerConfiguration.ts
@@ -10,8 +10,8 @@ export let isDebugEnabled = (): boolean => {
     return log.getLevel() === log.levels.DEBUG;
 };
 
-export let logDebug = (s: string | Error): void => {
+/*export let logDebug = (s: string | Error): void => {
     if (isDebugEnabled()) {
         log.debug(s);
     }
-};
+};*/
